refactor : path 패키지 분리
