+++ src/main/java/nextstep/subway/applicaion/path/ShortestDurationPathFinder.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion;
+package nextstep.subway.applicaion.path;
 
 import java.util.Collection;
 import java.util.List;
